+++ src/main/java/com/imsweb/naaccrxml/gui/pages/DictionariesPage.java
@@ -9,14 +9,21 @@ import java.awt.Component;
 import java.awt.Dimension;
 import java.awt.FlowLayout;
 import java.awt.Insets;
+import java.io.BufferedWriter;
+import java.io.File;
+import java.io.FileWriter;
 import java.io.IOException;
+import java.util.Comparator;
 import java.util.Vector;
 import java.util.regex.PatternSyntaxException;
 
 import javax.swing.Box;
+import javax.swing.JButton;
 import javax.swing.JComboBox;
 import javax.swing.JComponent;
+import javax.swing.JFileChooser;
 import javax.swing.JLabel;
+import javax.swing.JOptionPane;
 import javax.swing.JPanel;
 import javax.swing.JScrollPane;
 import javax.swing.JTabbedPane;
@@ -63,8 +70,8 @@ public class DictionariesPage extends AbstractPage {
 
         Vector<NaaccrDictionaryWrapper> standardDictionaries = new Vector<>();
         for (String version : NaaccrFormat.getSupportedVersions()) {
-            standardDictionaries.add(new NaaccrDictionaryWrapper(NaaccrXmlDictionaryUtils.getBaseDictionaryByVersion(version)));
-            standardDictionaries.add(new NaaccrDictionaryWrapper(NaaccrXmlDictionaryUtils.getDefaultUserDictionaryByVersion(version)));
+            standardDictionaries.add(new NaaccrDictionaryWrapper(NaaccrXmlDictionaryUtils.getBaseDictionaryByVersion(version), true));
+            standardDictionaries.add(new NaaccrDictionaryWrapper(NaaccrXmlDictionaryUtils.getDefaultUserDictionaryByVersion(version), false));
         }
         standardDictionaries.sort((o1, o2) -> o2.getDictionary().getNaaccrVersion().compareTo(o1.getDictionary().getNaaccrVersion()));
 
@@ -75,6 +82,63 @@ public class DictionariesPage extends AbstractPage {
         controlsPnl.add(Box.createHorizontalStrut(10));
         final JComboBox selectionBox = new JComboBox<>(standardDictionaries);
         controlsPnl.add(selectionBox);
+        controlsPnl.add(Box.createHorizontalStrut(25));
+        JButton extractBtn = new JButton("Extract to CSV...");
+        extractBtn.addActionListener(e -> {
+            JFileChooser fileChooser = new JFileChooser();
+            fileChooser.setFileSelectionMode(JFileChooser.FILES_ONLY);
+            fileChooser.setDialogTitle("Select Target File");
+            fileChooser.setApproveButtonToolTipText("Create CSV");
+            fileChooser.setMultiSelectionEnabled(false);
+            NaaccrDictionaryWrapper dictionary = (NaaccrDictionaryWrapper)selectionBox.getSelectedItem();
+            @SuppressWarnings("ConstantConditions")
+            String prefix = dictionary.isBase() ? "naaccr-dictionary-" : "user-defined-naaccr-dictionary-";
+            fileChooser.setSelectedFile(new File(fileChooser.getCurrentDirectory(), prefix + dictionary.getDictionary().getNaaccrVersion() + ".csv"));
+            if (fileChooser.showDialog(DictionariesPage.this, "Create CSV") == JFileChooser.APPROVE_OPTION) {
+                File targetFile = fileChooser.getSelectedFile();
+                if (targetFile.exists()) {
+                    int result = JOptionPane.showConfirmDialog(DictionariesPage.this, "Target file already exists, are you sure you want to replace it?", "Confirmation",
+                            JOptionPane.YES_NO_OPTION);
+                    if (result != JOptionPane.YES_OPTION)
+                        return;
+                }
+                try (BufferedWriter writer = new BufferedWriter(new FileWriter(targetFile))) {
+                    writer.write("NAACCR XML ID,NAACCR Number,Name,Start Column,Length,Record Types,Parent XML Element,Data Type");
+                    writer.newLine();
+                    dictionary.getDictionary().getItems().stream()
+                            .sorted(Comparator.comparing(NaaccrDictionaryItem::getNaaccrId))
+                            .forEach(item -> {
+                                try {
+                                    writer.write(item.getNaaccrId());
+                                    writer.write(",");
+                                    writer.write(item.getNaaccrNum() == null ? "" : item.getNaaccrNum().toString());
+                                    writer.write(",\"");
+                                    writer.write(item.getNaaccrName() == null ? "" : item.getNaaccrName());
+                                    writer.write("\",");
+                                    writer.write(item.getStartColumn() == null ? "" : item.getStartColumn().toString());
+                                    writer.write(",");
+                                    writer.write(item.getLength().toString());
+                                    writer.write(",\"");
+                                    writer.write(item.getRecordTypes() == null ? "" : item.getRecordTypes());
+                                    writer.write("\",");
+                                    writer.write(item.getParentXmlElement() == null ? "" : item.getParentXmlElement());
+                                    writer.write(",");
+                                    writer.write(item.getDataType() == null ? NaaccrXmlDictionaryUtils.NAACCR_DATA_TYPE_TEXT : item.getDataType());
+                                    writer.newLine();
+                                }
+                                catch (IOException | RuntimeException ex1) {
+                                    throw new RuntimeException(ex1); // doing that to make sure the loop is broken...
+                                }
+                            });
+                }
+                catch (IOException | RuntimeException ex2) {
+                    String msg = "Unexpected error creating CSV file\n\n" + ex2.getMessage();
+                    JOptionPane.showMessageDialog(DictionariesPage.this, msg, "Error", JOptionPane.ERROR_MESSAGE);
+                }
+                JOptionPane.showMessageDialog(DictionariesPage.this, "Extract successfully created!", "Success", JOptionPane.INFORMATION_MESSAGE);
+            }
+        });
+        controlsPnl.add(extractBtn);
 
         JPanel centerPnl = new JPanel(new BorderLayout());
         this.add(centerPnl, BorderLayout.CENTER);
@@ -198,12 +262,11 @@ public class DictionariesPage extends AbstractPage {
         columns.add("ID");
         columns.add("Num");
         columns.add("Name");
-        columns.add("Start Column");
+        columns.add("Start Col");
         columns.add("Length");
         columns.add("Record Types");
         columns.add("Parent XML Element");
         columns.add("Data Type");
-        columns.add("Validation Regex");
         columns.add("Padding");
         columns.add("Trimming");
 
@@ -218,7 +281,6 @@ public class DictionariesPage extends AbstractPage {
             row.add(item.getRecordTypes());
             row.add(item.getParentXmlElement());
             row.add(item.getDataType() == null ? NaaccrXmlDictionaryUtils.NAACCR_DATA_TYPE_TEXT : item.getDataType());
-            row.add(item.getRegexValidation());
             row.add(item.getPadding() == null ? NaaccrXmlDictionaryUtils.NAACCR_PADDING_RIGHT_BLANK : item.getPadding());
             row.add(item.getTrim() == null ? NaaccrXmlDictionaryUtils.NAACCR_TRIM_ALL : item.getTrim());
             rows.add(row);
@@ -228,10 +290,10 @@ public class DictionariesPage extends AbstractPage {
         _filterFld.setText("");
 
         _itemsTbl.getColumnModel().getColumn(1).setPreferredWidth(45); // item number
-        _itemsTbl.getColumnModel().getColumn(3).setPreferredWidth(30); // start column
+        _itemsTbl.getColumnModel().getColumn(3).setPreferredWidth(45); // start column
         _itemsTbl.getColumnModel().getColumn(4).setPreferredWidth(30); // length
         _itemsTbl.getColumnModel().getColumn(7).setPreferredWidth(40); // data type
-        _itemsTbl.getColumnModel().getColumn(10).setPreferredWidth(40); // trim
+        _itemsTbl.getColumnModel().getColumn(9).setPreferredWidth(40); // trim
 
         try {
             if (dictionary.getDictionaryUri().contains("user-defined"))
@@ -250,12 +312,19 @@ public class DictionariesPage extends AbstractPage {
 
         private NaaccrDictionary _dictionary;
 
-        public NaaccrDictionaryWrapper(NaaccrDictionary dictionary) {
+        private boolean _isBase;
+
+        public NaaccrDictionaryWrapper(NaaccrDictionary dictionary, boolean isBase) {
             _dictionary = dictionary;
+            _isBase = isBase;
         }
 
         public NaaccrDictionary getDictionary() {
             return _dictionary;
+        }
+
+        public boolean isBase() {
+            return _isBase;
         }
 
         @Override