+++ CHANGELOG.md
@@ -7,6 +7,7 @@
  - Changed type of causeOfDeath in all base dictionaries from digits to text.
  - Fixed start column of grouped item extendOfDisease10Dig in all base dictionaries.
  - Now display the number of Patient and Tumor tags found in the source XML file in the GUI.
+ - Added an option to the GUI for extracting the dictionaries as CSV files.
 
 **Version 3.0**